update fusion models after code restructure
