+++ FusionTransformer/models/late_fusion.py
@@ -1,13 +1,6 @@
-import torch
 import torch.nn as nn
-import torch.nn.functional as F
-import timm
 from FusionTransformer.models.spvcnn import SPVCNN
-from torchsparse.sparse_tensor import SparseTensor
-from FusionTransformer.models.STN import STN
-from FusionTransformer.models.transformer2D import Net2DSeg
-
-import numpy as np
+from FusionTransformer.models.image_models import Net2DSeg
 
 class Net3DSeg(nn.Module):
     def __init__(self,
@@ -55,101 +48,3 @@ class LateFusionTransformer(nn.Module):
         preds_lidar = self.lidar_backbone(data_dict["lidar"])
         out = {**preds_image, **preds_lidar}
         return out
-
-def test_Net2DSeg():
-    # 2D
-    batch_size = 1
-    img_width = 1200
-    img_height = 384
-
-    # 3D
-    num_coords = 2000
-    num_classes = 11
-
-    # 2D
-    img = torch.rand(batch_size, 3, img_height, img_width)
-    u = torch.randint(high=img_height, size=(batch_size, num_coords // batch_size, 1))
-    v = torch.randint(high=img_width, size=(batch_size, num_coords // batch_size, 1))
-    img_indices = torch.cat([u, v], 2)
-
-    # to cuda
-    img = img.cuda()
-    img_indices = img_indices.cuda()
-
-    net_2d = Net2DSeg(num_classes,
-                      dual_head=True,
-                      backbone_2d_kwargs={"block_number": 11})
-
-    net_2d.cuda()
-    out_dict = net_2d({
-        'img': img,
-        'img_indices': img_indices,
-    })
-    for k, v in out_dict.items():
-        print('Net2DSeg:', k, v.shape)
-
-
-def test_Net3DSeg():
-    in_channels = 4
-    num_coords = 2000
-    full_scale = 4096
-    num_seg_classes = 11
-
-    coords = torch.randint(high=full_scale, size=(num_coords, 3))
-    feats = torch.rand(num_coords, in_channels)
-    x  = SparseTensor(feats=feats, coords=coords).cuda()
-
-
-    net_3d = Net3DSeg(num_seg_classes,
-                      dual_head=True,
-                      backbone_3d_kwargs={'pres': 1, 'vres': 1})
-
-    net_3d.cuda()
-    out_dict = net_3d(x)
-    for k, v in out_dict.items():
-        print('Net3DSeg:', k, v.shape)
-
-
-def test_LateFusion():
-    # 2D
-    batch_size = 1
-    img_width = 1200
-    img_height = 384
-
-    # 3D
-    num_coords = 2000
-
-    # 2D
-    img = torch.rand(batch_size, 3, img_height, img_width)
-    u = torch.randint(high=img_height, size=(batch_size, num_coords // batch_size, 1))
-    v = torch.randint(high=img_width, size=(batch_size, num_coords // batch_size, 1))
-    img_indices = torch.cat([u, v], 2)
-    print(img_indices.shape)
-    # to cuda
-    img = img.cuda()
-
-    in_channels = 4
-    num_coords = 2000
-    full_scale = 4096
-    num_seg_classes = 11
-
-    coords = torch.randint(high=full_scale, size=(num_coords, 3))
-    feats = torch.rand(num_coords, in_channels)
-    x  = SparseTensor(feats=feats, coords=coords).cuda()
-
-    model = LateFusionTransformer(
-        num_class=num_seg_classes,
-        dual_head=True,
-        backbone_2d_kwargs={"block_number": 11},
-        backbone_3d_kwargs={'pres': 1, 'vres': 1}
-    )
-
-    model.cuda()
-    out_dict = model({"img": img, "img_indices": img_indices, "lidar": x})
-
-    for k, v in out_dict.items():
-        print('Fusion:', k, v.shape)
-
-if __name__ == '__main__':
-    # test_Net2DSeg()
-    test_LateFusion()
